Begins #2484: Remove usages of `org.apache.commons.lang3.tuple.Pair` et al (#2652)
This begins #2484 by introducing new `Pair` implementations to serve as alternatives to the one from Apache comments. It then uses the tests in a few places that do not have visible consequences. For example, we have a few places where we expose the old `Pair` class in our API (usually as a return value, but sometimes as a parameter). To fully remove the class, we'll need to replace those, but for now, this just adjusts usages that are entirely internal (which means mostly test classes for now).