Fixes #2449: LazyCloseableTest.testOpensLazilyExactlyOnceThreaded intermittently fails with: collection size <8> was less than <10> (#2657)
This updates `LazyCloseableTest.testOpensLazilyExactlyOnceThreaded` (and its equivalent test in `LazyOpenerTest`) so that they no longer rely on `Suppliers.parallel` to schedule tasks in multiple threads. Instead, it uses a new helper method which will spawn one thread for each `Supplier`. The helper method also uses a `CountDownLatch` to ensure that we do not start to execute tasks until all of the threads have been spawned. This ensures that there are as many concurrent accesses to the `LazyOpener`/`LazyCloseable` as possible during the test, though Java's thread scheduling algorithm remains non-deterministic.

This fixes #2449.