Resolves #2680: Allow RecordSerializers to validate serialization
There is a new method added to `RecordSerializer`, `validateSerialization`. This method takes both a `Message` and a serialized record, and it validates that calling `deserialize` on the bytes returns the original `Message`. In addition, there's a new configuration option on the `TransformedRecordSerializer` that allows it to be configured to validate a sample of serialized records (that sample configurable to "always" or "never") as this serializer has a somewhat more complicated story surrounding what it does to the bytes.

This resolves #2680.