Resolves #2689: Decompressing deserializer does not validate length of decompressed data
This adds a check to the `decompress` method to validate that the decompressed data coming from a transformed record serializer exactly matches the size encoded in the serialized data frame. This ensures that if there is some kind of error that results in the compressed blob either deserializing into something bigger or smaller than the original size, we notice and get a error right away (rather than, say, a protobuf deserialization error).

This resolves #2689.