Resolves #2691: Compressing serializer does not include checksum
This adds an additional call to ensure that the checksum gets included in the compressed data. This also adds an additional test that makes sure that if any of the compressed bits are flipped, we either get an error during decompression or we get back the original record (if, e.g., a bit is flipped that does not change data).

This resolves #2691.