Update RL dependency to 3.4.494.0 (REL-707)
This takes in a number of updates since Record Layer 3.4.490.0: https://github.com/FoundationDB/fdb-record-layer/blob/main/docs/ReleaseNotes.md#344940

This includes the fixes for  (Inflating corrupt compressed records) to more closely validate deserialized data as well as a number of Lucene fixes.

Other than the version number bump in `libraries.gradle`, the rest of the changes are from the rewrite tool.