([QA] "Database Error(42F59): could not find some of the provided index names") (REL-708)
The bitset calculation for some index types is currently off, in that it can include some indexes that are modeled in the Relational schema template. That includes synthetic index types (which is affecting Customer), but I think it can also include multi-type indexes, which may be what's going on for a few other databases. This updates the logic so that the index set that it is comparing against is more complete.