Fixes #2705: Permuted index can mis-assign fields in IndexKeyValueToPartialRecord
The calculation for the assigning index entries to fields used by the permuted min and max index wasn't quite right. There's a function that takes the fields ordered by their original place in the index (per-permutation) and then constructs a field extractor, meaning it needs to find the position in the entry accounting for the permutation. This worked by:

1. For grouping columns that precede the permuted aggregate, it preserved the index from the original expression. This is correct, as they are unpermuted.
1. For the permuted aggregate, it skips over `permutedCount` columns. This is correct, because the aggregate is moved from the end `permutedColumns` to the left.
1. For grouping columns that follow the permuted aggregate, it was _adding_ the `permutedCount`. This is _incorrect_, as those columns are moved `groupedCount` columns to the right, where `groupedCount` is equal to the number of aggregated columns. (Note that this value is actually independent from `permutedCount`.) So this should add `groupedCount` instead of `permutedCount`

In the existing tests that used this part of the planner, we only ever had one (1) permuted column and one (1) aggregate column, so the error wasn't hit by our existing tests. I've added a test where `groupedCount` > `permutedCount`, and sure enough, this exposed the issue.

This fixes #2705.