Fixes #2678: Default collation function depends on JVM setting for locale (#2679)
If unspecified, collation `FunctionKeyExpression`s now set their locale to `Locale.ROOT` instead of the machine's default Locale. This makes the contents of index entries more resilient to users changing their machine's default values. This does mean that users who are running with an incompatible locale will need to update their index definitions to reference the appropriate locale or will need to rebuild any indexes that don't already set a locale.

This fixes #2678. This is a specific instance of #2672.