Fixes #2715: Error message for stores with missing header can be misleading (#2716)
The previous logic used to decide what error message to present based on the `storeExistenceCheck`. This is incorrect, because it means we would report that the store would have no record data if the existence check supported opening stores with no header and no recods, but we should instead only report that if the store actually has no records. So now the error message is based on the data we actually find.

This fixes #2715.