Resolves #2733: Remove String.format calls
This removes as many `String.format` calls as could be done. For the most part, these were just using `%s` or `%d`, which can be replaced with String concatenation. For error messages, an effort was made to move the interpolated data into log message keys where appropriate so that the error messages are static strings (as much as possible), which is useful when trying to group together related errors in a production environment.

This resolves #2733.