Fixes #2741: Inverted range from tests that run non-overlapping parallel builds (#2742)
This updates the logic within the index builder tests so that it uses `BigInteger`s when splitting up the range. This is necessary because we can sometimes get large enough ranges between the `start` and `end` that intermediate steps may not be handled using 64-bit precision. These tests may be unnecessary as we mature the other parallel index build infrastructure and tests, but for now, this should unblock those hitting these flaky tests.

This fixes #2741.