Fixes #2750: Some calls to toUpperCase and toLowerCase are missing fixed Locale
This updates a few calls to `toUpperCase()` and `toLowerCase()` to specify `Locale.ROOT` instead of `Locale.getDefault()`. This fixes an issue where different handling of `"i".toUpperCase()` or `"I".toLowerCase()` were resulting in the `ArithmeticValue` to not successfully encapsulate multiplication. For the record, it turns out we're already using a PMD rule that forces the user to specify a `Locale`, but it only checks that a locale was specified, not that the `Locale` is a specific value. Also, the PMD checks don't run on test code, which is why those classes in test logic were not specifying a locale at all.

This fixes #2750.