Fixes build with debugBuild=false (REL-734)
There was an issue with how planHash and explain commands could interact. The plan hash operation can end up inserting `explain ${query}` into the plan cache. This means that when we go to execute the command for the `explain`, we can read the plan from the cache, which means that we could get the non-debug plan. This messes up our asserts, so to fix this, we now run both operations with the debugger.