Fixes #2813: Fail fast when Batch GRV Rate Limit Exceeded
Loading the metadata version key does not seem to throw
Batch GRV rate limit exceeded, while other operations on the same
transaciton do. Instead it stalls forever, and the tests end up
timing out at the junit level.
This change introduces a new combine that will fail when the loading
the record store state fails, instead of waiting forever for the
metadata version to be loaded.