Combine #2674 and #2812 (#2815)
* Resolves #2663: Add query support for indexed artithmetic functions

This adds a new `FunctionKeyExpression` representing some basic arithmetic functions on `long`s, including `sum` and `bitand` (useful for when long fields are used as bitsets, and the user wants to define an index on individual bits in the bitset). This includes adding support for expressing the expression as a `Value` so that it can be matched within the Cascades planner. From the first test fixture added, it appears that the planner can already handle basic queries on this field, though it doesn't always recognize when it can skip a record resolution (e.g., when the function value is in both the result set and the index or when the argument fields are indexed).

This resolves #2663.

Includes #2812 and https://github.com/alecgrieser/fdb-record-layer/pull/1 to match constant values with literal values.
---------

Co-authored-by: Normen Seemann <nseemann@apple.com>
Co-authored-by: normen662 <60754965+normen662@users.noreply.github.com>