([SQL] Support bitwise operator predicates/indexes) (REL-737)
* Update Relational index generator and vistors for indexed arithmetic functions

This adds support for the bitwise functions within the index generator. It also adds the bitwise functions to the function catalog so that this can be used directly. This requires https://github.com/FoundationDB/fdb-record-layer/pull/2674 for the `FunctionKeyExpression` to be available for Relational to be defined on.

Cannot currently add tests of aggregate indexes using the DDL. It also avoids stripping literals in bit expressions

---------

Co-authored-by: Youssef Hatem <y_hatem@apple.com>