Switch to Dropwizard metrics-core managed by shared dependencies BOM (REL-746)
* Switch to Dropwizard metrics-core managed by shared dependencies BOM

This updates our dependency of codahale metrics-core to one that is managed by the shared dependencies BOM. This updates our version from 3.0.2 to 4.2.26, and so it allows us to make use of a no-op metric registry that was previously unavailable. The old no-op implementation has been hollowed out, and now it just contains a static `INSTANCE` method to avoid re-creating objects.

* Remove reference to GlobalInterceptors

The class has been moved to `ConfiguratorRegistry` in more modern versions of gRPC. There's a comment that this was about preventing ClassNotFound errors during runtime, which I'm not seeing anymore when I run locally with the line removed.