Log fdbcli status after failed test
This is an attempt to investigate Issue #2842 in a remote environment.
The theory is that the cluster is in a state where it is rejecting
all transactions, but the hope is that it is in a state where
status works, and will tell us what is wrong.

This is resiliant to not having `fdbcli` on your path or if you
are using `fdb-environment.properties` instead of the default
cluster file, but it will just print errors that it could not get
the status.