([SQL] PlanCache fails with RowStructs in prepared statement parameter) (REL-760)
* add test showing failing type promotion for tuple list

*  ([SQL] PlanCache fails with RowStructs in prepared statement parameter)

This updates the way that array structs are handled within the AstNormalizer so that they get messaged at that point into DynamicMessages. Previously, the `Struct`s were left as is, which meant that they were treated as an `ARRAY(ANY)`. The plan cache, meanwhile, already had `ARRAY(RECORD(...))` as the expected type, as the structs were translated into dynamic messages elsewhere when planning and executing. This aligns the types in the cache with the types handed to the cache to prevent errors when using cached plans with these kinds of queries.