Fixes #2890: Prevent timeouts in chaosMergeAndUpdateTest
There were two issues::
1. PR #2887 caused us to start taking and using a timer snapshot, but
snapshots cannot be used if the timer was reset. The test here used
the same timer for the thread saving records, and the thread doing
merges, and when the test commits the transaction, it resets the
timer. Thus, when the insert thread committed, it would reset the
transaction and cause the merge to fail. So, now, the merge call
creates a new timer instead of reusing the one from the class. If
this is ever needed, it can become an optional parameter to
`explicitMergeIndex`
2. The insert thread would call mergeQueue.put(false) when it finished,
but when the merge thread failed, it would stop popping off the
queue, and subsequently the insert thread would wait forever, and
not be joinable. Thus, the test would timeout. So, now, the merger
records the failure, but continues to process merges, so it will
eventually allow the `put` at the end of the inserter thread.