Fixes #2901: In-unions on aggregate indexes can be planned with insufficient comparison keys in the presence of repeated fields (#2902)
The logic to calculate the `Ordering` of aggregate queries has been augmented to include all values, not just the values even when there are repeated fields. This should be acceptable because with aggregate indexes, repeated fields don't change the cardinality in the same way that they do with `Value` indexes. (The value index logic may also consider tweaking how it works a bit, but that's not something this change attempts to handle.) The test suite for the permuted min/max indexes have been augmented to include more tests with repeated keys, including ones that would produce results that get incorrectly de-duplicated with previous `Ordering` logic.

This fixes #2901.