(Update Relational to JDK 17) (REL-799)
This upgrades the code base to use tooling that is compatible with Java 17. This includes updating our build toolchain versions to values that are compatible with Java 17, and then responding to the fall out.

It doesn't update our build to use Java 17 (yet), though that should be a pretty straightforward change if we want to do that now.