Fixes #2936: Missing subspace provider in FDBRecordStore logs
There were a handful of places in the `FDBRecordStore` class that were were logging but didn't include the subspace provider as context. This means that we couldn't tell what store the logs were trying to log about, which made it hard to correlate with other operations to the store. This resolves that by adding in the missing information to the missing logs.

This fixes #2936.