Fixes #2941: In-union plans don't consider sargability of in-value predicates
This updates the in-union planning in the old planner so that it behaves like in-join planning. That is to say, it attempts to use a base plan that only has extracted IN-predicates that are pushed into a scan. Non-sargable predicates are passed as is to the filter plan. For backwards compatibility reasons, this is hidden behind a planner configuration property.

This fixes #2941.