Some work to stabilize
This removes the exceptions, but now, it will occasionally just
have an infinite loop when trying to close.
I think this is due to the fact that ArrayDeque is not thread safe,
and it potentially adds and removes from the queue at the same time,
causing it to get into a state where calling poll returns null even
though it's not empty, and it doesn't actually remove the element.