Fixes #548: Fix concurrency issues with MapPipelinedCursor.close
The underlying issue is that multiple threads are interacting with
the (non-thread-safe) queue at the same time, in complex ways.
We could add `synchronized` to a `close` and `onNext`, (or some
part of onNext), but instead I changed the cleanup around so that
it only interacts with the pipeline in a single chain of futures.