Fix #2953: FlatMapPipelinedCursor can fail with NoSuchElementException or NullPointerException if concurrently closed
This updates the `FlatMapPipelinedCursor` to control all of its accesses of the (non-thread safe) pipeline via synchronized methods. It furthermore tries
to ensure that any cursors that get created are closed. The use of `synchronized` methods here, unlike in the `MapPipelinedCursro`, is necessitated by the
possibility that the user stops iterating on a cursor and calls `close()`. If nothing calls `.onNext()` on that closed cursor, then nothing will end up closing
the generated cursors in the alternative architecture.

This fixes #2953.