refactoring, addressing further comments.
- Massive renaming, now we have `TempTableScanExpression`, `TempTableInsertExpression`, respective `TempTableInsertPlan`, and `TempTableScanPlan`, PB names are also consistent with that.
- Moved `PEnumLightValue`, `PUUID`, `PFDBRecordVersion`, `PComparableObject`, `PQueryResult`, and `PTempTable` to `record_metadata.proto` since these are runtime objects.
Removed the name from `TempTable`.
Use `Value` to “refer” to a `TempTable` in the `EvaluationContext` instead of a plain `CorrelationIdentifier`, in all `TempTableScanExpression`, `TempTableInsertExpression`, respective `TempTableInsertPlan`, and `TempTableScanPlan`. Made the expressions’ constructors private and added two factory methods that enable the customer to create an expression using a constant binding, or a correlated binding.
- Fixed the `translateCorrelations` implementations in `UpdateExpression`, `InsertExpression`, and `DeleteExpression`. Now it rebuilds the expression using the correct translated inner.