Fixes #2971: Test failure in MoreAsyncUtilTest.getWithDeadlineRunsOnExecutor
There was a race condition in a test assertion, and it was actually possible for a test callback to land on the same thread as the test worker if setting up the future chain
took longer than the deadline in the test. That should be fairly unlikely, but it could happen, and so this expands the assertion to include that case.

This fixes #2971.