Fixes #2975: Test failure in MoreAsyncUtilTest.executeDelayedCallbackOnExecutor (#2976)
As was done in #2972, this updates the assertions for some tests in `MoreAsyncUtilTest` that were designed to make sure that the wrong thread pool wasn't accidentally being used for callbacks on delayed futures. The test neglected the fact that the calling thread could also end up executing the callback if the future completed quickly (or if there were a delay between when the future is created and the callback added). This can happen in certain environments, including during some of our nightly CI builds, though it's not the only place it has been noticed.

This fixes #2975.