Fixes #2986: TimeWindowLeaderboardIndex throws NPE on null ranks
This modifies the `TimeWindowRankAndEntry` struct to allow the returned `rank` to be empty. This represents the case where a record was not found in one of the leaderboards and so a `null` rank is returned. Note that this returns us to the state of the world prior to #2834, which is what introduced the struct (with its non-null `rank` field). Prior to that PR, the rank could be `null`, and so any code that worked prior to that commit (which is also the most recent change as of writing) should be able to be trivially adapted to this new version.

This fixes #2986.