AnalyzerChooser no longer depends on text
The only implementation that actually referenced the text was in
tests.
This gets us on a path to an index having a fixed analyzer (at least
for the duration of the transaction), which can simplify the
FDBDirectoryWrapper.
The premise for adding AnalyzerChooser was to use a different
analyzer depending on the text to better support different languages,
but this seems problematic, because you won't know when searching what
analyzer was used. i.e. if you have a document that is mixed,
and then search only for one language, it will be confused.
Instead a more complicated analyzer should be used to support this,
if needed.