LuceneAnalyzerCombinationProvider can store analyzers in fields
Since there is no longer a text parameter, it can generater the
wrappers once. This may result in some wasted CPU generating unused
analyzers (e.g. when not querying or not writing), but it should be
compensated by the fact that it consistently creates 2 wrappers,
rather than creating a wrapper for every record update, or every
query.