Runnable relational server jar (REL-831)
* Change fdb-relational-server to produce a jar with a manifest & dependencies

This allows you to take the resulting output and do something like:
```
java -jar fdb-relational-server/.out/libs/fdb-relational-server-2433B351-SNAPSHOT.jar
```
To start a relational server.
This does not pickup `fdb-environment.properties` but continues to
get those from the environment. For this reason, I added logging
for these two important environment variables.

* Use shadowJar plugin to generate shadowJar instead of using jar

The previous attempt by adding it to the jar mostly worked, but
I ran into trouble with it sometimes not being able to find
`DefaultPromise$1`.
Since we are already using shadowjar elsewhere in the project, I
switched to that, which seems to work well.

* Include shadow jar in .dist and publishing

I confirmed that on running `build` it now includes the shadow
jar in the .dist folder. I have not confirmed the actual publishing,
but it doesn't complain about compilation.

I noticed the jdbc doesn't do this, so I made the same change there
so it now adds the driver shadowJar to the .dist, and should
publish it.

* Remove explicit calls to shadow in build scripts

Both jdbc and the server now generate the shadow jar as part of
the standard build.

* Guard logger statements

PMD caught the newly added log files, and I added the other check
because I noticed it.