Download old version of the server for mixed mode testing  (REL-837)
* Got code to download a specific version

I spent a lot of time trying to fight with declaring these as
dependencies, and using something like:
```
from configurations.mixedModeTests
```
But it refused to use anything other than the SNAPSHOT from the
project itself.

* Move download to yaml-tests and update the version

It now downloads with the `all` classifier, meaning that it will
download the shadow jar, which should be runnable all on it's own.

I tried for most of a day to get it to download the latest
non-SNAPSHOT release, but I couldn't, and left some comments about
my findings.

* Put the downloaded jar in the build directory & send path to tests

Plus some naming cleanup

* Clean the external server directory before downloading

This way the only jar in that directory will be the one we want to
use