Strict continuation assertions (REL-844)
* QueryCommand now fails with a clear error if there is no contiuation

If the yamsql command expects a continuation and more results,
QueryCommand will fail with a clearer error. Before, it would loop
back to the beginning and then, most likely, fail because the result
set doesn't match.

* QueryCommand now fails if you don't expect all the results, across continuations

The bulk is going through and expecting the remainder of the results,
including empty results sets after the last one when it is an
even multiple.
We may at some point want to have an explicit command to ignore
the remainder, but I didn't feel it was justified yet.

* Disable JDBCYamlIntegrationTests.aggregateIndexTests due to continuations

Because handling of continuations is now required, this test
fails for JDBC because that does not support continuations