Yamsql test supported_version
This adds a bunch of tests that supported_version causes the right
queries to be run or not run as appropriate.

I tried to do this without separate files, and without actually
involving the rest of the relational layer, but I couldn't figure
out how to do that without having to write a lot of stub methods.
I think figuring that out would probably be valuable, but I opted
to not do it today.

Note: When we automate changing from !current_version to a release
we'll want to skip over this entire resources directory.