Add support to skip a test from the config side
Not everything is supported JDBC yet, so we need the ability to
skip some tests.
It seemed clearer to do it from the config side, so you can easily
see what is not supported for a config. The configs that are not
supported for a test will be visible when you run the test.