Add a quickTest target to make it easier to run with just the embedded
A lot of times (I assume) people will want to test changes to the
relational core, or record-layer core, and will want faster tests
then running all of the configs. This will allow easily doing so
in intelliJ with a new `quickTest` target when you got o run a
test.
The quickTest target is not run as part of `./gradlew build`, it
has to be explicitly requested, so we won't get redundant test runs
in our CI.