Improve LuceneIndexTestDataModel (#3076)
* Cleanup LuceneIndexTestDataModel start parameter
The saveRecords methods had a `start` parameter that would
influence the timestamp. Originally this was added to make the
records saved look more like a timestamp. This probably isn't
necessary, but certainly it can be a field, cleaning up the api
that everyone found confusing.

* Test that adding old documents can create new partitions
* Add support for isSynthetic to LuceneIndexTestDataModel.recordsUnderTest
* Additional test that deleting some records affects lucene partition info
This is testing to make sure that if we delete every other record
(sorted by partition key), then the partition metadata gets updated
correctly, and everything is consistent.

This is pretty well tested elsewhere, but this adds a variant that
uses `LuceneIndexTestDataModel`

* Remove unused import