Rework YamlIntegrationTests to not use inheritance (#3083)
At a high level, this introduces an `@YamlTest` annotation,
and configurations that will run for every `@TestTemplate`.
This allows us to more easily run with a variety of combinations.

* Introduce YamlTest annonation & extension
This just adds one config, and copies one test using it.

* Change YamlConnectionFactory to not unwrap
This fixes a spotbugs error. MultiServerConnectionFactory still
has a problem of not closing, in theory, and that will probably
need to be resolved too, as that gets brought into "production"
code.

* Add new JDBCInProcessConfig
This is basically the same as JDBCInProcessYamlIntegrationTests
except we don't wrap the exception in the afterAll

* Extract ExternalServer logic out of junit Extension
Just moving the code into a different class

* Add MultiServerConfig
This required running two servers with separate ports.
I could have had them try to share an external server, but this
seems simpler, especially if we later have configs that run with
extra versions of the server.

* Switch YamlIntegrationTests to use new system
This allows deleting all the other implementations.
YamlTestBase still exists as it is used for the supported_version
tests

* Add a quickTest target to make it easier to run with just the embedded

A lot of times (I assume) people will want to test changes to the
relational core, or record-layer core, and will want faster tests
then running all of the configs. This will allow easily doing so
in intelliJ with a new `quickTest` target when you got o run a
test.
The quickTest target is not run as part of `./gradlew build`, it
has to be explicitly requested, so we won't get redundant test runs
in our CI.

* Change SupportedVersionTest to not extend YamlTestBase

* Add support for excluding a YamlTestConfig from the test itself

This allows other projects to use @YamlTest exclude some of their
tests from some configs.

* Remove RunExternalServerExtension
now unused

* Change MultiServerConnectionFactoryTest to return an mock
We now call unwrap, so that needs to work. We don't call anything
else though, which makes Mockito great.

* Added @ExcludedTestConfig.reason
I added a reason to explain why the test is excluded

* Add back configurations to utilitize FORCE_CONTINUATIONS
This adds a new config to run with FORCE_CONTINUATIONS=true,
and runs it with both combinations of mixed mode.
This required reworking @ExcludeYamlTestConfig to support the new
permutations.