Increment version in gradle.properties automatically (#3082)
* Update relase workflow to accept artifact version

This updates the release workflow so that it runs when someone manually triggers it rather than on every commit. It also takes an argument for the version, which hopefully will fix some tests that are failing during the current build.

It still doesn't update the build to do any of the other things we associate with releases (e.g., updating any of the values storing the current version), but it may give us insight into the release build itself.

* Add script to increment version automatically and hook it into the release

* increment version during release

* update patch release build to merge into main at the end

* remove artifact version task

* update workflows based on testing

* Add configuration to publish to GitHub packages and to Maven Central

This adds the gradle configuration necessary to publish to maven central and GitHub packages if requested. It also then calls those methods during the release step of our GitHub actions.

* separate out FDB and env setup to avoid running setup twice

* verison bump

* some cleanup

* update to 4.1 ; add documentation about maven central

* update release notes for #3111

* remove compiler.xml which is no longer necessary (as stated by the .gitignore) but will now update with every release

* update versionutils.py to make more internally consistent