Fixes #3139: Plan Hash Instability in recursive union query
This updates the plan hashes for some of the plan operators associated with recursive CTEs. They previously hashed the quantifier directly, which could include things like enum hashes, which are not stable across JVMs. This resulted in multi-server test failures. This updates the plan hash calculations to be based off of the child _plans_ instead of the child _quantifiers_, which have stable plan hashes.

To test this, I published this change to my maven local repo. I then configured the yaml tester to use that newly published library against my embedded driver, and I validated that the failing tests in `recursive-cte.yamsql` now passed. In order to make this actually mergeable, though, I needed to update the `supported_version` on those two tests to `!current_version`, which means we won't actually get coverage of this until the next release. We could consider adding additional configurations that compare embedded versus external on the same version, which would help cover things like this, but I didn't add that in this PR.

I did update `recursive-cte.yamsql` with query plans, which made it clearer what plans to look at and felt like good things to have. I've checked in the now necessary metrics files as well for those plans.

This fixes #3139.