Fix broken build in Main (#3143)
Two tests now fail on Main. The tests had a `supported_version` that was
enabled once a release was cut but a plan hash change now make them fail
when they are run on multi-server config.
This PR reintroduces the `!current_version` for the tests such that
these will not run in multi-server mode