Resolves #1288: Automatically publish to maven central
This should complete our maven central automation. It configures artifact signing (required for maven central upload), and it makes use of the nexus publishing plugin to upload artifacts to maven central to a staging repository, close the staging repository (which validates the expected invariants from maven central), and then release the artifacts it just uploaded. Once the final step has happened, the new artifacts should be available for maven central download (modulo the time it takes for CDNs to be updated).

This resolves #1288.