Kill all yamsql external servers before starting tests
When debugging the tests, if you stop it sends SIGKILL, which kills
the test process, but not any of the child processes. This means
that when you restart the tests they'll fail, because the ports are
taken. We could have it dynamically generate ports, but that would
most likely lead to developers having hundreds of servers running.
Instead, this code, at startup will find all running servers and
kill them.