Add gradle command to do mixed-mode tests against many versions
This allows running the mixed-mode tests against a bunch of versions.
It is currently set to the 50 most up-to-date versions (based on
version ordering, not publish date).
That being said, builds older than 4.0.559.0 did not publish a
server, so we can't test against anything older than that.
I chose the 50 pretty arbitrarily, we can adjust as time goes on.
I wanted to also check the `Build-Date` but it appears that is not
getting correctly filled in.