Resolves #3155: FRL respects PLAN_CACHE_*_MAX_ENTRIES
I discovered that running a lot of YamlTestConfigs against the
same Embedded server can result in exhausting a heap of 512MB.
The main cause (at least 300mb) was the plan cache, but there
was no way to control the size, so this passes those options
along when running `FRL`.
`TransactionBoundEmbeddedRelationalEngine` already passes these
options along.

A limit of 10 should cover any tests we have, and any short-term
parallesim, or tests with multiple schema templates.