Add support for prepared statements in JDBC (#3116)
This PR enables support for prepared statements on the JDBC server.
Specifically, support for PreparedStatement `setObject()` and
`setArray()`.
Also, `JDBCRelationalConnection` needed support for array creation.
A slew of type conversion utilities were added in support of these
features.
In general, we have 3 models that we are dealing with:
- Java (including primitives, [], null)
- JDBC (including primitive, Array, Struct, NULL)
- Protobuf

And because of the way the API is structured, we may need to perform
conversions between all 3. Specifically, we support the following:
- Java and SQL primitives are treated the same way
- Java arrays are only supported in `Connection.createArrayOf`, where
the type of array elements is also provided
- SQL `Array` and `Struct` are assumed in all cases of nesting. That is,
an array of elements would be assumed to have SQL `Array` if there is
any nested array. The reasoning is that we would need the additional
type information in these cases
Some cleanup and restructuring of that code may be warranted, in
addition to null, error and edge case testing.

The protobuf for the `Array` definition was changed in order to support
the extra type information.
The change was made to be backwards compatible at the protobuf level but
the old server (before this change) cannot handle arrays in parameters.
In order to facilitate multi-version testing against the older servers
we may be able to cherry-pick these commits to the server's branch.