Give up on getting Coverage comment in pull request itself. (#3162)
The `madrapps/jacoco-report` plugin does not have a way of setting the
base sha for the `workflow_run` case and so won't get the proper diffs.
Instead put it inline in the `pull_request` workflow (which does not
have access to add PR comments) and there just write to job summary.