Resolves #3167: RecordQueryPlanComplexityException includes the full plan string
This updates the plan string in the default `RecordQueryPlanComplexityException` to include only the first 1k characters in the plan string. This decreases the total size of the data pushed to the exception for logging. It also updates the `RecordQueryPlanComplexityException` class so that it extends `RecordQueryException` and uses the log message keys now available in order to convey information instead of relying updating the error message with each error.

This resolve #3167.