Add support for running YAML tests against a server built from local branch (#3161)
In order to increase coverage - specifically, running the current
version on multiple JVMs - this change adds the ability to build and run
multi-server tests against the current local version.
The test run will treat the locally-built server the same as any other
downloaded server. That means that 4 test runs will be performed:
- embedded then remote (with and without continuation forcing)
- remote then embedded (with and without continuation forcing)

This effectively adds one more downloaded server to the list.
An option was considered (and can be quite easily implemented) to reduce
coverage by treating this server differently (and only run a subset of
the configs) but given the large number of servers we already plan to
download anyway, this didn't seem material.