Run mixed mode tests in a workflow, and publish results to release notes (#3166)
This adds a workflow to run the yaml-tests at a given tag against the
last 50 releases, and adds the results to `docs/ReleaseNotes.md`.
Note: Currently it will only download back to `4.0.559.1`, because older
builds didn't have a server to do mixed-mode testing against.

Currently this is a manual workflow that someone would run after a
successful build, but once proven, we can change it to automatically run
after a release.

There is some discussion in @alecgrieser's review, but I will also point
out here that the results are somewhat weakened by the use of
`!supported_version`. For example, `prepared.yamsql` is set to
`!supported_version: 4.1.6.0` because that support only exists in
`4.1.6.0`, but it was also added to `4.0.559.6` (which was released
after the test run in the sample below). We also have issues where the
behavior may change between two versions. None of that is visible here,
and would have to be noted in the release notes, and found there. But
hopefully the fact that we run single version mixed-mode tests during
PRB and release will prevent issues from completely slipping through.

Sample output (an actual run, but the link is to a different workflow):

> #### Mixed Mode Test Results
> 
> Mixed mode testing run against the following previous versions:
> ✅`4.0.559.1`, ✅`4.0.559.2`, ✅`4.0.559.3`, ✅`4.0.559.4`, ❌`4.0.561.0`,
✅`4.0.562.0`, ✅`4.0.564.0`, ✅`4.0.565.0`, ✅`4.0.566.0`, ✅`4.0.567.0`,
✅`4.0.568.0`, ✅`4.0.569.0`, ✅`4.0.570.0`, ✅`4.0.571.0`, ✅`4.0.572.0`,
✅`4.0.573.0`, ✅`4.0.574.0`, ✅`4.0.575.0`, ✅`4.1.4.0`, ✅`4.1.5.0`,
✅`4.1.6.0`
> 
> [See full test
run](https://github.com/FoundationDB/fdb-record-layer/actions/runs/13331380671)